This fixes #2.
Generation of namcouple file is now placed in its own module create_namcouple.py
and executed earlier (before working directories of the models are created).

In this new module, a temporary working directory for the flux_calculator is created
and the executable is called with the argument --genrate_namcouple.

The created namcouple then overwrites the existing one in the input folder.
From there it is broadcasted to the working directories of the models (locally or globally)