This commit fixes #4
Script modify_cclm_restart_file.py uses laf...nc files to get a corrected land mask.
Since these files currently only exist for the beginning of the month,
the file could not be found when the run should start in the
middle of the month.

Fix: Do not take the actual starting date but rather the beginning of the
month of the starting date.

This solution might be improved in the future.