Worked on attempt handling.
If an attempt fails a new job is submitted to run the next attempt.
The evaluation is happening in the global work directory.
The option that the output is copied from the local work directories
to the output is now deprecated.
This path of the scripts was anyway not working.