Introduced input_name
If input_name is empty, there is only one input folder called "input" and
run_name is set according to the global_settings.py therein.

If input_name is not empty, it corresponds to one of the folder in "input".
The run_name is set to that input_name and any run_name setting in the
global_settings.py is ignored.